Image path revamp

+++ fotogalleri/backend/models.py
@@ -1,4 +1,4 @@
-from django.db import models
+from django.db.models import Model, CharField, ImageField, DateTimeField, ForeignKey, CASCADE
 from json import dumps, loads
 from os.path import join
 
@@ -7,19 +7,17 @@ def new_image_path(instance, filename):
     return filename
 
 
-class ImageMetadata(models.Model):
+class ImageMetadata(Model):
     '''
     Metadata model for uploaded images.
     Stores simple metadata such as file name, path, and time and date of upload.
-
-    Images are stored in the following manner /<year>/<event_name>/<image_name>
     '''
-    year = models.CharField(max_length=256, blank=False, null=False)
-    event = models.CharField(max_length=256, blank=False, null=False)
-    image = models.ImageField(upload_to=new_image_path)
-    upload_time = models.DateTimeField(auto_now_add=True, blank=True)
+    # FIXME: refactor on_delete
+    path = ForeignKey('ImagePath', on_delete=CASCADE, blank=True, null=True)
+    image = ImageField(upload_to=new_image_path)
+    upload_time = DateTimeField(auto_now_add=True, blank=True)
     # JSON array formatted as a string for containing all thumbnails
-    thumbnails_json = models.CharField(max_length=256, blank=False, null=False, default='[]')
+    thumbnails_json = CharField(max_length=256, blank=False, null=False, default='[]')
 
     def set_thumbnails(self, new_thumbnails):
         thumbnails = self.thumbnails + list(new_thumbnails)
@@ -39,13 +37,40 @@ class ImageMetadata(models.Model):
 
         if not is_saved and self.image:
             oldfile = self.image.name
-            newfile = join(self.year, self.event, oldfile)
-
-            self.image.storage.save(newfile, self.image)
-            self.image.name = newfile
-            self.image.close()
-            self.image.storage.delete(oldfile)
+            newfile = join(self.path.full_path if self.path else '', self.image.name)
+            if newfile != oldfile:
+                self.image.storage.save(newfile, self.image)
+                self.image.name = newfile
+                self.image.close()
+                self.image.storage.delete(oldfile)
 
     def __str__(self):
-        return '{path};{time}'.format(path=self.image,
-                                      time=self.upload_time)
+        return '{path};{time}'.format(path=self.image, time=self.upload_time)
+
+
+class ImagePath(Model):
+    # FIXME: refactor on_delete
+    parent = ForeignKey('self', on_delete=CASCADE, blank=True, null=True)
+    path = CharField(max_length=256, blank=False, null=False, editable=False, unique=True)
+
+    def _get_full_path(self):
+        return join(self.parent.full_path if self.parent else '', self.path)
+
+    full_path = property(_get_full_path)
+
+    def _is_root_child(self):
+        '''
+        Returns true if this ImagePath is a first descendant of the root (i.e. empty parent).
+        '''
+        return not self.parent
+
+    is_root_child = property(_is_root_child)
+
+    def __str__(self):
+        return self.full_path
+
+    @staticmethod
+    def create(path, parent):
+        new_image_path = ImagePath(path=path, parent=parent)
+        new_image_path.save()
+        return new_image_path
